Major improvement to setpath.vim plugin
Got rid of hardcoded values in script for project definitions. The script now look for a file called .vimproj in the folder tree (up to the root) and parse the file. This file describe the project.
The MakeProg option is available to configure a main make prog for the current project. This still needs work. I'm not totally happy with the handling of the make_prog at this point. But the management of files is functional is pretty neat! And I added a connection to FuzzyFinder pluggin. If fur is installed, you just need to map a command to the function VPFuzzyFinder.